refactor(utils.price_display): move some functions from utils to price_display.py
